requirements.txt
fastapi
uvicorn[standard]
openai
mem0ai
python-dotenv==1.0.1
supabase
vecs
pydantic
